refactor:게시글 삭제 권한을 작성자에게만 부여하고, 관리자는 게시글 공개 상태만 변경

+++ src/boards/boards.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { Board } from './boards.entity';
 import { BoardStatus } from './boards-status.enum';
 import { CreateBoardDto } from './dto/create-board.dto';
@@ -35,8 +35,11 @@ export class BoardsService {
 
     // // 특정 게시글 조회 기능
     async getBoardDetailById(id: number): Promise<Board> {
-        const foundBoard = await this.boardsRepository.findOneBy({ id: id }); // await 추가
-        if (!foundBoard) {
+        const foundBoard = await this.boardsRepository.createQueryBuilder('board')
+            .leftJoinAndSelect('board.user','user') //사용자 정보를 조인
+            .where('board.id = id', {id})
+            .getOne(); 
+        if(!foundBoard){
             throw new NotFoundException(`Board with ID ${id} not found`);
         }
         return foundBoard;
@@ -97,8 +100,12 @@ export class BoardsService {
     }   
 
     // // 게시글 삭제 기능  
-    async deleteBoardById(id: number): Promise<void> {  
+    async deleteBoardById(id: number,logginedUser:User): Promise<void> {  
         const foundBoard = await this.getBoardDetailById(id);
+        // 작성자와 요청한 사용자가 같은지 확인
+        if (foundBoard.user.id !== logginedUser.id){
+            throw new UnauthorizedException('Do not have permission to delete this board')
+        }
         await this.boardsRepository.delete(foundBoard);  
     }
